Move env file to storage MCP server
