+++ servers/cdn/env.ts
@@ -1,7 +1,6 @@
 import { z } from "zod";
 
 export const envSchema = z.object({
-  GROQ_API_KEY: z.string().min(1, "Missing groq api key"),
   BUNNY_STORAGE_ZONE_NAME: z.string().min(1, "Missing storage zone name"),
   BUNNY_STORAGE_ACCESS_KEY: z.string().min(1, "Missing storage access key"),
   BUNNY_PULLZONE_URL: z.string().url("Missing pullzone URL"),
